fix(server): use server leaf certificate to verify client signature
If the serverCertificate contains a chain, the signature calculation shall be done only with the leaf Certificate. For backward compatibility a Server shall check the signature with the full chain if the check with the leaf Certificate fails.
Hence the verification shall also use the leaf certificate and if failed the chain.

See also: OPC UA Part 4, V1.05, 5.7.3 Activate Session, Table 17 - ActivateSession Service Parameters