fix(client): use client leaf certificate to verify server signature
This parameter is calculated by appending the clientNonce to the clientCertificate and signing the resulting sequence of bytes.

If the clientCertificate contains a chain, the signature calculation shall be done only with the leaf Certificate. For backward compatibility a Client shall check the signature with the full chain if the check with the leaf Certificate fails.

See also: OPC UA Part 4, V1.05, 5.7.2 Create Session, Table 15 - CreateSession Service Parameters